Optimize genotype initialization and add phenotype conversion

+++ evolution/utils.py
@@ -4,4 +4,4 @@ import numpy as np
 
 def read_image(image_path: str) -> np.ndarray:
     image = Image.open(image_path)
-    return np.asarray(image)
+    return np.asarray(image, dtype=float)
